feat: rename package suparest to db
