+++ pkg/db/aggregate.go
@@ -1,4 +1,4 @@
-package suparest
+package db
 
 import (
 	"strconv"
