+++ pkg/db/delete.go
@@ -1,4 +1,4 @@
-package suparest
+package db
 
 import (
 	"github.com/valyala/fasthttp"
