+++ pkg/db/postgrest_request.go
@@ -1,4 +1,4 @@
-package suparest
+package db
 
 import (
 	"fmt"
