+++ pkg/db/update.go
@@ -1,4 +1,4 @@
-package suparest
+package db
 
 import (
 	"encoding/json"
